+++ python/examples/01_the_switch_and_the_light/switch.py
@@ -1,3 +1,4 @@
+from arete_sdk.client import Client
 import atexit
 import ssl
 import sys
@@ -6,8 +7,6 @@ import RPi.GPIO as GPIO
 from threading import Thread
 from time import sleep
 
-sys.path.append('../../src/arete_sdk')
-from client import Client
 
 GPIO04 = 4